🐛 Bug : 로그인 후 유저 정보를 가져오지 못하는 오류
회원가입 및 로그인을 구현하는 방식을 route handler에서 server action으로 변경
